[S10P22A108-206] Feat: add Querying Diaries, Modifying Diary, Delete Diary
Former-commit-id: 59f4131093078a3bb719ceae8a96d9cdf4beeb68

+++ back-end/src/main/java/com/nightletter/domain/member/repository/MemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.nightletter.domain.member.entity.Member;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberCustomRepository{
+public interface MemberRepository extends JpaRepository<Member, Integer>, MemberCustomRepository{
 
 	Member findMemberByOAuth2Id(String OAuth2Id);
